Review hardening: usize->uint64 parity, blame parse strictness, test coverage, script fixes
Whole-solution review (CliSupport/Diff/Git/Jj) found no serious bugs. Changes: (1) widen the usize-mapped integers to uint64 for exact Rust parity now that the API is free to change - DiffStat (FilesChanged/Insertions/Deletions), Hunk (Old/New Start/Lines), BranchStatus (Ahead/Behind/TrackedChanges/Untracked/Conflicts), and Git.RevListCount; (2) git blame author-time now parses with AllowLeadingSign (no surrounding whitespace, matching Rust parse::<i64>()); (3) tests added for git blame --line-porcelain, the numeric-vs-lexicographic Version ordering case, Credential.ToString redaction, and the forced jj --color never argv contract; (4) check-env.ps1/.sh no longer point at a non-existent scripts/init.* - they now name the real build step (dotnet tool restore + build).

+++ src/VcsToolkit.Diff/Model.fs
@@ -6,15 +6,15 @@ namespace VcsToolkit.Diff
 type DiffStat =
     {
         /// Number of files changed.
-        FilesChanged: int
+        FilesChanged: uint64
         /// Lines added (`insertions(+)`).
-        Insertions: int
+        Insertions: uint64
         /// Lines removed (`deletions(-)`).
-        Deletions: int
+        Deletions: uint64
     }
 
     /// Build a `DiffStat`.
-    static member Create(filesChanged: int, insertions: int, deletions: int) =
+    static member Create(filesChanged: uint64, insertions: uint64, deletions: uint64) =
         { FilesChanged = filesChanged
           Insertions = insertions
           Deletions = deletions }
@@ -47,13 +47,13 @@ type DiffLine =
 type Hunk =
     {
         /// Start line in the old file (the `-<start>` of the `@@` header).
-        OldStart: int
+        OldStart: uint64
         /// Line count in the old file (defaults to 1 when the `,<count>` is omitted).
-        OldLines: int
+        OldLines: uint64
         /// Start line in the new file (the `+<start>` of the `@@` header).
-        NewStart: int
+        NewStart: uint64
         /// Line count in the new file (defaults to 1 when the `,<count>` is omitted).
-        NewLines: int
+        NewLines: uint64
         /// Text after the closing `@@` (the function/section heading); empty when none.
         Section: string
         /// The hunk body, one entry per `+`/`-`/` ` line.
